--- CodeStarterCamp_Week1/icecream.swift
@@ -1,22 +1,0 @@
-//
-//  main.swift
-//  CodeStarterCamp_Week1
-//
-//  Created by yagom.
-//  Copyright © yagom academy. All rights reserved.
-//
-
-import Foundation
-
-func makeIcecream(){
-    for _ in 0...7{
-        print("**********")
-    }
-}
-func drawStick(){
-    for _ in 0...3{
-        print("   | |")
-    }
-}
-makeIcedream()
-drawStick()
